+++ changelog.rst
@@ -10,6 +10,8 @@ ChangeLog
 * Add supports for ``Babyl``, ``MH`` and ``MMDF`` mailbox types.
 * Add new ``-f``/``--sources-format`` option to force the type of mail sources
   on opening.
+* Add new ``-u``/``--force-unlock`` option to force the removal of a lock on
+  mailboxes on opening.
 * Split-up the selection of mails candidate for removal and the deletion itself
   into two distinct batch operations.
 * Add issue templates to guide users to report bugs and request features.